requirements.txt
python-telegram-bot[asyncio]
fastapi
aiohttp
asyncpg
python-dotenv
uvicorn
nest_asyncio